+++ src/main/java/com/example/demo/repository/StudyRoomRepository.java
@@ -19,5 +19,5 @@ public interface StudyRoomRepository extends JpaRepository<StudyRoom, Long> {
     ArrayList<StudyRoom> findStudyRoomsByUniversityAndBuilding(String university, String building);
 
     @Query("SELECT o From Order o WHERE o.studyRoom = :studyRoom AND o.date = :date AND o.state != '반납' AND o.state != '취소' AND o.startTime BETWEEN :startTime AND :endTime")
-    List<Order> findByDateBetweenTime(@Param("studyRoom") StudyRoom studyRoom, @Param("date") LocalDate date, @Param("startTime") float startTime, @Param("endTime") float endTime);
+    List<Order> findByDateBetweenTime(@Param("studyRoom") StudyRoom studyRoom, @Param("date") LocalDate date, @Param("startTime") double startTime, @Param("endTime") double endTime);
 }
